Extract interfaces for repositories
