+++ app/src/main/java/net/squanchy/service/repository/SpeakerRepository.java
@@ -2,51 +2,15 @@ package net.squanchy.service.repository;
 
 import java.util.List;
 
-import net.squanchy.service.firebase.FirebaseDbService;
-import net.squanchy.service.firebase.model.FirebaseSpeaker;
 import net.squanchy.speaker.domain.view.Speaker;
-import net.squanchy.support.lang.Checksum;
-import net.squanchy.support.lang.Optional;
 
 import io.reactivex.Observable;
 
 import static net.squanchy.support.lang.Lists.map;
 
-public class SpeakerRepository {
+public interface SpeakerRepository {
 
-    private final FirebaseDbService dbService;
-    private final Checksum checksum;
+    Observable<List<Speaker>> speakers();
 
-    public SpeakerRepository(FirebaseDbService dbService, Checksum checksum) {
-        this.dbService = dbService;
-        this.checksum = checksum;
-    }
-
-    public Observable<List<Speaker>> speakers() {
-        return dbService.speakers()
-                .map(firebaseSpeaker -> firebaseSpeaker.getSpeakers())
-                .map(speakers -> map(speakers, this::toSpeaker));
-    }
-
-    public Observable<Speaker> speaker(String speakerId) {
-        return dbService.speakers()
-                .map(firebaseSpeakers -> firebaseSpeakers.getSpeakers())
-                .flatMap(Observable::fromIterable)
-                .filter(firebaseSpeaker -> firebaseSpeaker.getId().equals(speakerId))
-                .map(this::toSpeaker);
-    }
-
-    private Speaker toSpeaker(FirebaseSpeaker firebaseSpeaker) {
-        return Speaker.Companion.create(
-                firebaseSpeaker.getId(),
-                checksum.getChecksumOf(firebaseSpeaker.getId()),
-                firebaseSpeaker.getName(),
-                firebaseSpeaker.getBio(),
-                Optional.fromNullable(firebaseSpeaker.getCompany_name()),
-                Optional.fromNullable(firebaseSpeaker.getCompany_url()),
-                Optional.fromNullable(firebaseSpeaker.getPersonal_url()),
-                Optional.fromNullable(firebaseSpeaker.getPhoto_url()),
-                Optional.fromNullable(firebaseSpeaker.getTwitter_username())
-        );
-    }
+    Observable<Speaker> speaker(String speakerId);
 }
