+++ app/src/main/java/net/squanchy/service/repository/injection/RepositoryModule.java
@@ -2,8 +2,11 @@ package net.squanchy.service.repository.injection;
 
 import net.squanchy.service.firebase.FirebaseDbService;
 import net.squanchy.service.repository.EventRepository;
+import net.squanchy.service.repository.firebase.EventRepositoryImpl;
 import net.squanchy.service.repository.SpeakerRepository;
 import net.squanchy.service.repository.VenueRepository;
+import net.squanchy.service.repository.firebase.SpeakerRepositoryImpl;
+import net.squanchy.service.repository.firebase.VenueRepositoryImpl;
 import net.squanchy.support.lang.Checksum;
 
 import dagger.Module;
@@ -18,7 +21,7 @@ public class RepositoryModule {
             Checksum checksum,
             SpeakerRepository speakerRepository
     ) {
-        return new EventRepository(dbService, checksum, speakerRepository);
+        return new EventRepositoryImpl(dbService, checksum, speakerRepository);
     }
 
     @Provides
@@ -26,11 +29,11 @@ public class RepositoryModule {
             FirebaseDbService dbService,
             Checksum checksum
     ) {
-        return new SpeakerRepository(dbService, checksum);
+        return new SpeakerRepositoryImpl(dbService, checksum);
     }
 
     @Provides
     VenueRepository venueRepositoryrRepository(FirebaseDbService dbService) {
-        return new VenueRepository(dbService);
+        return new VenueRepositoryImpl(dbService);
     }
 }
